# openapi.yaml
openapi: 3.1.0
info:
  title: TIC API
  version: "0.1.0-mvp"
  description: >
    Contract for the TIC MVP backend. Enum values in this file must match
    shared/constants.json exactly — see docs/CONTRACT_CHANGE_PROCESS.md
    before editing either file.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [access_token]
                properties:
                  access_token: { type: string }
        "401":
          description: Refresh token invalid or expired
  /auth/register:
    post:
      summary: First-admin signup — always creates a brand-new organization and an ADMIN user in it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_name, name, email, password]
              properties:
                org_name: { type: string }
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "201":
          description: Organization + admin created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "409":
          description: Email already registered
...
  /agent/register:
    post:
      summary: Register a desktop agent/device for the authenticated employee
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_name, os, agent_version]
              properties:
                device_name: { type: string }
                os: { type: string, enum: [WINDOWS, MACOS, LINUX] }
                agent_version: { type: string }
      responses:
        "200":
          description: Device registered
          content:
            application/json:
              schema:
                type: object
                required: [device_id, status]
                properties:
                  device_id: { type: integer }
                  status: { type: string }
  /activity/batch:
    post:
      summary: Ingest an aggregated activity batch from the desktop agent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./backend/app/schemas/activity_batch.schema.json"
      responses:
        "200":
          description: Batch accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityBatchResponse" }
        "422":
          description: Batch failed schema validation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        user:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
            role: { type: string, enum: [EMPLOYEE, MANAGER, ADMIN] }
    ActivityBatchResponse:
      type: object
      required: [status, batch_id]
      properties:
        status: { type: string, enum: [accepted] }
        batch_id: { type: string }
        processed:
          type: object
          properties:
            application_sessions: { type: integer }
            browser_tab_sessions: { type: integer }
            context_switches: { type: integer }
...